refactor authentication classes
