+++ football-stats-predictions/src/main/kotlin/com/footballdata/football_stats_predictions/config/SecurityConfig.kt
@@ -1,4 +1,4 @@
-package com.footballdata.football_stats_predictions
+package com.footballdata.football_stats_predictions.config
 
 import com.footballdata.football_stats_predictions.repositories.UserRepository
 import com.footballdata.football_stats_predictions.service.JwtUserDetailsService
@@ -74,4 +74,3 @@ class SecurityConfig {
     @Bean
     fun encoder(): PasswordEncoder = BCryptPasswordEncoder()
 }
-
