+++ football-stats-predictions/src/main/kotlin/com/footballdata/football_stats_predictions/dto/AuthenticationResponse.kt
@@ -1,4 +1,4 @@
-package com.footballdata.football_stats_predictions.webservice
+package com.footballdata.football_stats_predictions.dto
 
 data class AuthenticationResponse(
     val accessToken: String,
