+++ football-stats-predictions/src/main/kotlin/com/footballdata/football_stats_predictions/dto/RegisterRequest.kt
@@ -1,0 +1,6 @@
+package com.footballdata.football_stats_predictions.dto
+
+data class RegisterRequest(
+    val username: String,
+    val password: String
+)
