+++ football-stats-predictions/src/main/kotlin/com/footballdata/football_stats_predictions/dto/TokenResponse.kt
@@ -1,0 +1,5 @@
+package com.footballdata.football_stats_predictions.dto
+
+data class TokenResponse(
+    val token: String
+)
