--- football-stats-predictions/src/main/kotlin/com/footballdata/football_stats_predictions/webservice/RefreshTokenRequest.kt
@@ -1,5 +1,0 @@
-package com.footballdata.football_stats_predictions.webservice
-
-data class RefreshTokenRequest(
-    val token: String
-)
